First draft for helical acquisition for 7-BM.
Also use some code from 2-BM for dealing with scans with missing angles.
