+++ tomoScanApp/Db/tomoScan_Helical_settings.req
@@ -1,0 +1,10 @@
+# This file is used by autosave to determine what PVs to save
+# It is also used by tomoScan to determine what PVs to create
+
+# This is the file for the tomoscan_helical derived class.
+
+##########################
+# Helical Scan information
+##########################
+$(P)$(R)PixelsPer360Deg
+$(P)$(R)HelicalMisalignment
